fix CustomEvent initialization
